Copy IPAddress-related types from CoreFX-master
