+++ src/System.Net.Primitives/src/System/Net/IPAddress.cs
@@ -15,12 +15,11 @@ namespace System.Net
     ///     Provides an Internet Protocol (IP) address.
     ///   </para>
     /// </devdoc>
-    [Serializable]
     public class IPAddress
     {
-        public static readonly IPAddress Any = new IPAddress(0x0000000000000000);
-        public static readonly IPAddress Loopback = new IPAddress(0x000000000100007F);
-        public static readonly IPAddress Broadcast = new IPAddress(0x00000000FFFFFFFF);
+        public static readonly IPAddress Any = new ReadOnlyIPAddress(0x0000000000000000);
+        public static readonly IPAddress Loopback = new ReadOnlyIPAddress(0x000000000100007F);
+        public static readonly IPAddress Broadcast = new ReadOnlyIPAddress(0x00000000FFFFFFFF);
         public static readonly IPAddress None = Broadcast;
 
         internal const long LoopbackMask = 0x00000000000000FF;
@@ -51,19 +50,6 @@ namespace System.Net
         /// </summary>
         private int _hashCode;
 
-        // Maximum length of address literals (potentially including a port number)
-        // generated by any address-to-string conversion routine.  This length can
-        // be used when declaring buffers used with getnameinfo, WSAAddressToString,
-        // inet_ntoa, etc.  We just provide one define, rather than one per api,
-        // to avoid confusion.
-        //
-        // The totals are derived from the following data:
-        //  15: IPv4 address
-        //  45: IPv6 address including embedded IPv4 address
-        //  11: Scope Id
-        //   2: Brackets around IPv6 address when port is present
-        //   6: Port (including colon)
-        //   1: Terminating null byte
         internal const int NumberOfLabels = IPAddressParserStatics.IPv6AddressBytes / 2;
 
         private bool IsIPv4
@@ -76,7 +62,7 @@ namespace System.Net
             get { return _numbers != null; }
         }
 
-        internal uint PrivateAddress
+        private uint PrivateAddress
         {
             get
             {
@@ -516,7 +502,7 @@ namespace System.Net
             }
         }
 
-        [Obsolete("This property has been deprecated. It is address family dependent. Please use IPAddress.Equals method to perform comparisons. http://go.microsoft.com/fwlink/?linkid=14202")]
+        [Obsolete("This property has been deprecated. It is address family dependent. Please use IPAddress.Equals method to perform comparisons. https://go.microsoft.com/fwlink/?linkid=14202")]
         public long Address
         {
             get
@@ -546,6 +532,10 @@ namespace System.Net
                 {
                     if (PrivateAddress != value)
                     {
+                        if (this is ReadOnlyIPAddress)
+                        {
+                            throw new SocketException(SocketError.OperationNotSupported);
+                        }
                         PrivateAddress = unchecked((uint)value);
                     }
                 }
@@ -609,8 +599,8 @@ namespace System.Net
             int hashCode;
             if (IsIPv6)
             {
-                const int addressAndScopeIdLength = IPAddressParserStatics.IPv6AddressBytes + sizeof(uint);
-                Span<byte> addressAndScopeIdSpan = stackalloc byte[addressAndScopeIdLength];
+                const int AddressAndScopeIdLength = IPAddressParserStatics.IPv6AddressBytes + sizeof(uint);
+                Span<byte> addressAndScopeIdSpan = stackalloc byte[AddressAndScopeIdLength];
 
                 MemoryMarshal.AsBytes(new ReadOnlySpan<ushort>(_numbers)).CopyTo(addressAndScopeIdSpan);
                 Span<byte> scopeIdSpan = addressAndScopeIdSpan.Slice(IPAddressParserStatics.IPv6AddressBytes);
@@ -623,12 +613,9 @@ namespace System.Net
             }
             else
             {
-                Span<uint> addressOrScopeIdSpan = stackalloc uint[1];
-                addressOrScopeIdSpan[0] = _addressOrScopeId;
- 
                 // For IPv4 addresses, we use Marvin on the integer representation of the Address.
                 hashCode = Marvin.ComputeHash32(
-                    MemoryMarshal.AsBytes(addressOrScopeIdSpan),
+                    MemoryMarshal.AsBytes(MemoryMarshal.CreateReadOnlySpan(ref _addressOrScopeId, 1)),
                     Marvin.DefaultSeed);
             }
 
@@ -671,7 +658,12 @@ namespace System.Net
             return new IPAddress(address);
         }
 
-        [MethodImpl(MethodImplOptions.NoInlining)]
         private static byte[] ThrowAddressNullException() => throw new ArgumentNullException("address");
+
+        private sealed class ReadOnlyIPAddress : IPAddress
+        {
+            public ReadOnlyIPAddress(long newAddress) : base(newAddress)
+            { }
+        }
     }
 }
