+++ src/System.Private.Uri/src/System/IPv6AddressHelper.cs
@@ -3,8 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
-using System.Text;
-using System.Globalization;
+using System.Diagnostics;
 
 namespace System
 {
@@ -12,108 +11,111 @@ namespace System
     // The idea is to stay with static helper methods and strings
     internal static partial class IPv6AddressHelper
     {
-        // fields
-#if !MONO
-        private const int NumberOfLabels = 8;
-#endif
-        // Lower case hex, no leading zeros
-        private const string CanonicalNumberFormat = "{0:x}";
-        private const string EmbeddedIPv4Format = ":{0:d}.{1:d}.{2:d}.{3:d}";
-        private const char Separator = ':';
-
         // methods
 
-        internal static string ParseCanonicalName(string str, int start, ref bool isLoopback, ref string scopeId)
+        internal static unsafe string ParseCanonicalName(string str, int start, ref bool isLoopback, ref string scopeId)
         {
-            unsafe
-            {
-                ushort* numbers = stackalloc ushort[NumberOfLabels];
-                // optimized zeroing of 8 shorts = 2 longs
-                ((long*)numbers)[0] = 0L;
-                ((long*)numbers)[1] = 0L;
-                isLoopback = Parse(str, numbers, start, ref scopeId);
-                return '[' + CreateCanonicalName(numbers) + ']';
-            }
-        }
+            ushort* numbersPtr = stackalloc ushort[NumberOfLabels];
+            // optimized zeroing of 8 shorts = 2 longs
+            ((long*)numbersPtr)[0] = 0L;
+            ((long*)numbersPtr)[1] = 0L;
+            Span<ushort> numbers = new Span<ushort>(numbersPtr, NumberOfLabels);
+            Parse(str, numbersPtr, start, ref scopeId);
+            isLoopback = IsLoopback(numbers);
 
-        internal static unsafe string CreateCanonicalName(ushort* numbers)
-        {
             // RFC 5952 Sections 4 & 5 - Compressed, lower case, with possible embedded IPv4 addresses.
 
             // Start to finish, inclusive.  <-1, -1> for no compression
-            KeyValuePair<int, int> range = FindCompressionRange(numbers);
+            (int rangeStart, int rangeEnd) = FindCompressionRange(numbers);
             bool ipv4Embedded = ShouldHaveIpv4Embedded(numbers);
 
-            StringBuilder builder = new StringBuilder();
+            Span<char> stackSpace = stackalloc char[48]; // large enough for any IPv6 string, including brackets
+            stackSpace[0] = '[';
+            int pos = 1;
+            int charsWritten;
+            bool success;
             for (int i = 0; i < NumberOfLabels; i++)
             {
                 if (ipv4Embedded && i == (NumberOfLabels - 2))
                 {
+                    stackSpace[pos++] = ':';
+                    
                     // Write the remaining digits as an IPv4 address
-                    builder.AppendFormat(CultureInfo.InvariantCulture, EmbeddedIPv4Format,
-                        numbers[i] >> 8, numbers[i] & 0xFF, numbers[i + 1] >> 8, numbers[i + 1] & 0xFF);
+                    success = (numbers[i] >> 8).TryFormat(stackSpace.Slice(pos), out charsWritten);
+                    Debug.Assert(success);
+                    pos += charsWritten;
+
+                    stackSpace[pos++] = '.';
+                    success = (numbers[i] & 0xFF).TryFormat(stackSpace.Slice(pos), out charsWritten);
+                    Debug.Assert(success);
+                    pos += charsWritten;
+
+                    stackSpace[pos++] = '.';
+                    success = (numbers[i + 1] >> 8).TryFormat(stackSpace.Slice(pos), out charsWritten);
+                    Debug.Assert(success);
+                    pos += charsWritten;
+
+                    stackSpace[pos++] = '.';
+                    success = (numbers[i + 1] & 0xFF).TryFormat(stackSpace.Slice(pos), out charsWritten);
+                    Debug.Assert(success);
+                    pos += charsWritten;
                     break;
                 }
 
                 // Compression; 1::1, ::1, 1::
-                if (range.Key == i)
-                { // Start compression, add :
-                    builder.Append(Separator);
+                if (rangeStart == i)
+                {
+                    // Start compression, add :
+                    stackSpace[pos++] = ':';
                 }
-                if (range.Key <= i && range.Value == (NumberOfLabels - 1))
-                { // Remainder compressed; 1::
-                    builder.Append(Separator);
+
+                if (rangeStart <= i && rangeEnd == NumberOfLabels)
+                {
+                    // Remainder compressed; 1::
+                    stackSpace[pos++] = ':';
                     break;
                 }
-                if (range.Key <= i && i <= range.Value)
+
+                if (rangeStart <= i && i < rangeEnd)
                 {
                     continue; // Compressed
                 }
 
                 if (i != 0)
                 {
-                    builder.Append(Separator);
+                    stackSpace[pos++] = ':';
                 }
-                builder.AppendFormat(CultureInfo.InvariantCulture, CanonicalNumberFormat, numbers[i]);
+                success = numbers[i].TryFormat(stackSpace.Slice(pos), out charsWritten, format: "x");
+                Debug.Assert(success);
+                pos += charsWritten;
             }
 
-            return builder.ToString();
+            stackSpace[pos++] = ']';
+            return new string(stackSpace.Slice(0, pos));
         }
 
-        // RFC 5952 Section 4.2.3
-        // Longest consecutive sequence of zero segments, minimum 2.
-        // On equal, first sequence wins.
-        // <-1, -1> for no compression.
-        private static unsafe KeyValuePair<int, int> FindCompressionRange(ushort* numbers)
+        private static unsafe bool IsLoopback(ReadOnlySpan<ushort> numbers)
         {
-            int longestSequenceLength = 0;
-            int longestSequenceStart = -1;
+            //
+            // is the address loopback? Loopback is defined as one of:
+            //
+            //  0:0:0:0:0:0:0:1
+            //  0:0:0:0:0:0:127.0.0.1       == 0:0:0:0:0:0:7F00:0001
+            //  0:0:0:0:0:FFFF:127.0.0.1    == 0:0:0:0:0:FFFF:7F00:0001
+            //
 
-            int currentSequenceLength = 0;
-            for (int i = 0; i < NumberOfLabels; i++)
-            {
-                if (numbers[i] == 0)
-                { // In a sequence
-                    currentSequenceLength++;
-                    if (currentSequenceLength > longestSequenceLength)
-                    {
-                        longestSequenceLength = currentSequenceLength;
-                        longestSequenceStart = i - currentSequenceLength + 1;
-                    }
-                }
-                else
-                {
-                    currentSequenceLength = 0;
-                }
-            }
-
-            if (longestSequenceLength >= 2)
-            {
-                return new KeyValuePair<int, int>(longestSequenceStart,
-                    longestSequenceStart + longestSequenceLength - 1);
-            }
-
-            return new KeyValuePair<int, int>(-1, -1); // No compression
+            return ((numbers[0] == 0)
+                            && (numbers[1] == 0)
+                            && (numbers[2] == 0)
+                            && (numbers[3] == 0)
+                            && (numbers[4] == 0))
+                           && (((numbers[5] == 0)
+                                && (numbers[6] == 0)
+                                && (numbers[7] == 1))
+                               || (((numbers[6] == 0x7F00)
+                                    && (numbers[7] == 0x0001))
+                                   && ((numbers[5] == 0)
+                                       || (numbers[5] == 0xFFFF))));
         }
 
         // Returns true if the IPv6 address should be formated with an embedded IPv4 address:
@@ -350,211 +352,6 @@ namespace System
         internal static unsafe bool IsValid(char* name, int start, ref int end)
         {
             return InternalIsValid(name, start, ref end, false);
-        }
-
-        //
-        // Parse
-        //
-        //  Convert this IPv6 address into a sequence of 8 16-bit numbers
-        //
-        // Inputs:
-        //  <member>    Name
-        //      The validated IPv6 address
-        //
-        // Outputs:
-        //  <member>    numbers
-        //      Array filled in with the numbers in the IPv6 groups
-        //
-        //  <member>    PrefixLength
-        //      Set to the number after the prefix separator (/) if found
-        //
-        // Assumes:
-        //  <Name> has been validated and contains only hex digits in groups of
-        //  16-bit numbers, the characters ':' and '/', and a possible IPv4
-        //  address
-        //
-        // Returns:
-        //  true if this is a loopback, false otherwise. There is no failure indication as the sting must be a valid one.
-        //
-        // Throws:
-        //  Nothing
-        //
-
-        internal static unsafe bool Parse(string address, ushort* numbers, int start, ref string scopeId)
-        {
-            int number = 0;
-            int index = 0;
-            int compressorIndex = -1;
-            bool numberIsValid = true;
-
-            //This used to be a class instance member but have not been used so far
-            int PrefixLength = 0;
-            if (address[start] == '[')
-            {
-                ++start;
-            }
-
-            for (int i = start; i < address.Length && address[i] != ']';)
-            {
-                switch (address[i])
-                {
-                    case '%':
-                        if (numberIsValid)
-                        {
-                            numbers[index++] = (ushort)number;
-                            numberIsValid = false;
-                        }
-
-                        start = i;
-                        for (++i; address[i] != ']' && address[i] != '/'; ++i)
-                        {
-                            ;
-                        }
-                        scopeId = address.Substring(start, i - start);
-                        // ignore prefix if any
-                        for (; address[i] != ']'; ++i)
-                        {
-                            ;
-                        }
-                        break;
-
-                    case ':':
-                        numbers[index++] = (ushort)number;
-                        number = 0;
-                        ++i;
-                        if (address[i] == ':')
-                        {
-                            compressorIndex = index;
-                            ++i;
-                        }
-                        else if ((compressorIndex < 0) && (index < 6))
-                        {
-                            //
-                            // no point checking for IPv4 address if we don't
-                            // have a compressor or we haven't seen 6 16-bit
-                            // numbers yet
-                            //
-
-                            break;
-                        }
-
-                        //
-                        // check to see if the upcoming number is really an IPv4
-                        // address. If it is, convert it to 2 ushort numbers
-                        //
-
-                        for (int j = i; (address[j] != ']') &&
-                                        (address[j] != ':') &&
-                                        (address[j] != '%') &&
-                                        (address[j] != '/') &&
-                                        (j < i + 4); ++j)
-                        {
-                            if (address[j] == '.')
-                            {
-                                //
-                                // we have an IPv4 address. Find the end of it:
-                                // we know that since we have a valid IPv6
-                                // address, the only things that will terminate
-                                // the IPv4 address are the prefix delimiter '/'
-                                // or the end-of-string (which we conveniently
-                                // delimited with ']')
-                                //
-
-                                while ((address[j] != ']') && (address[j] != '/') && (address[j] != '%'))
-                                {
-                                    ++j;
-                                }
-                                number = IPv4AddressHelper.ParseHostNumber(address, i, j);
-                                unchecked
-                                {
-                                    numbers[index++] = (ushort)(number >> 16);
-                                    numbers[index++] = (ushort)number;
-                                }
-                                i = j;
-
-                                //
-                                // set this to avoid adding another number to
-                                // the array if there's a prefix
-                                //
-
-                                number = 0;
-                                numberIsValid = false;
-                                break;
-                            }
-                        }
-                        break;
-
-                    case '/':
-                        if (numberIsValid)
-                        {
-                            numbers[index++] = (ushort)number;
-                            numberIsValid = false;
-                        }
-
-                        //
-                        // since we have a valid IPv6 address string, the prefix
-                        // length is the last token in the string
-                        //
-
-                        for (++i; address[i] != ']'; ++i)
-                        {
-                            PrefixLength = PrefixLength * 10 + (address[i] - '0');
-                        }
-                        break;
-
-                    default:
-                        number = number * 16 + Uri.FromHex(address[i++]);
-                        break;
-                }
-            }
-
-            //
-            // add number to the array if its not the prefix length or part of
-            // an IPv4 address that's already been handled
-            //
-
-            if (numberIsValid)
-            {
-                numbers[index++] = (ushort)number;
-            }
-
-            //
-            // if we had a compressor sequence ("::") then we need to expand the
-            // numbers array
-            //
-
-            if (compressorIndex > 0)
-            {
-                int toIndex = NumberOfLabels - 1;
-                int fromIndex = index - 1;
-
-                for (int i = index - compressorIndex; i > 0; --i)
-                {
-                    numbers[toIndex--] = numbers[fromIndex];
-                    numbers[fromIndex--] = 0;
-                }
-            }
-
-            //
-            // is the address loopback? Loopback is defined as one of:
-            //
-            //  0:0:0:0:0:0:0:1
-            //  0:0:0:0:0:0:127.0.0.1       == 0:0:0:0:0:0:7F00:0001
-            //  0:0:0:0:0:FFFF:127.0.0.1    == 0:0:0:0:0:FFFF:7F00:0001
-            //
-
-            return ((numbers[0] == 0)
-                            && (numbers[1] == 0)
-                            && (numbers[2] == 0)
-                            && (numbers[3] == 0)
-                            && (numbers[4] == 0))
-                           && (((numbers[5] == 0)
-                                && (numbers[6] == 0)
-                                && (numbers[7] == 1))
-                               || (((numbers[6] == 0x7F00)
-                                    && (numbers[7] == 0x0001))
-                                   && ((numbers[5] == 0)
-                                       || (numbers[5] == 0xFFFF))));
         }
     }
 }
